Inlined the model for the DCC2775 test case inside the test class
