--- tests/Doctrine/Tests/Models/DDC2775/AdminRole.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Doctrine\Tests\Models\DDC2775;
-
-/** @Entity */
-class AdminRole extends Role
-{
-}
